issue#11/support-for-git-tags
Upgraded the manage.py build command such that it now accepts the tag
flag. If requested, it looks up the tag and the commit from github. It
fails if github does not have this tag.

Relevant tests were added to test_commands.py, and the stub data
included.

This also closes JIRA Issue AWS-103.