AWS-141 Spin up and tear down testing environment
Working towards having a cluster being spun up.

DockerRunners for solid services, such as Consul, and Postgres,
have been included. These services now use a .ready() and a
.running(). One that says if the container is alive, and the
other that the service is ready. This means arbritrary waits
are no longer required, and that the provisioning should not break.

The base tests for the manage scripts have been added, but the task
for spin up has not been fully fleshed out yet.