AWS-141 Spin up and tear down cluster
Modified the consul healthcheck to use a request, as the consulate
API does not seem to crash if consul is unresponsive, consistently.

Upgraded vagrant provisioning to include a psql client, required
for the provisioning of the postgres databases.

Upgraded the timer, such that it can exit based on the required
return value. It is only used for True/False, but can be used for
any desire.

Relevant tests of the above were upgraded.