AWS-141 Updates to provisioning behaviour
Removed the call back feature of the provisioner. The driving force for this change, is
that the provisioner should not need to look for relevant information based on the
container it is provisioning. Instead, it is cleaner to have the container request to
be provisioned using a certain class.

Tests were upgraded to include this style of provisioning.