AWS-141 spinup and teardown test cluster
Added a Test Provisioner and Test Runner. These two classes, as
named, are used to provision a set of tests with things such as,
ports, ids, etc., and then to run them.

Tests have been included for the above classes.

Tasks that run tests in the cluster and stop the test cluster have
been added.

Tests for these tasks have been added.

This should close AWS-141, as the core features have been implemented.
The next pull request should be used to actually run the tests, and fine
missing gaps.