AWS-187 Test environment to work with ADSRex
This should be the last commit for this branch. Any following commits
will be to fix tests or typos before being merged into master. The
following describes the changes:

Vagrant file - Updated to allow more memory for Solr, otherwise the heapsize defined in
the docker container is not sufficient

utils - added a yaml function that loads yaml contents into an ordered
dictionary. Tests included

Some services will ignore unknown services, such as Solr, which does not
care if the service requested does not have any configuration file.

Single task included to spin up, run tests, teardown. Tests included. YAML
file is used for the setup at the moment.

A lot of the stub data was updated to ensure the tests that were being
used in adsrex passed correctly.

Extra stub data (consul +/- postgres) added for the following services:
 - citation_helper
 - export
 - orcid
 - solr
 - vis
 - myads
 - orcid
 - recommender

Updated the central README docs, and added to the README docs for each
provisioner, detailing the instructions of how to include stub data.